+++ packages/lib/package/components/index.ts
@@ -27,10 +27,12 @@ import { default as MazRadioButtons } from './MazRadioButtons.vue'
 import { default as MazSelect } from './MazSelect.vue'
 import { default as MazSlider } from './MazSlider.vue'
 import { default as MazSpinner } from './MazSpinner.vue'
+import { default as MazStepper } from './MazStepper.vue'
 import { default as MazSwitch } from './MazSwitch.vue'
 import { default as MazTabsBar } from './MazTabsBar.vue'
 import { default as MazTabsContent } from './MazTabsContent.vue'
 import { default as MazTabsContentItem } from './MazTabsContentItem.vue'
+import { default as MazTextarea } from './MazTextarea.vue'
 import { default as MazTransitionExpand } from './MazTransitionExpand.vue'
 
 export {
@@ -61,10 +63,12 @@ export {
   MazSelect,
   MazSlider,
   MazSpinner,
+  MazStepper,
   MazSwitch,
   MazTabsBar,
   MazTabsContent,
   MazTabsContentItem,
+  MazTextarea,
   MazTransitionExpand,
 }
 
@@ -96,9 +100,11 @@ export default {
   MazSelect,
   MazSlider,
   MazSpinner,
+  MazStepper,
   MazSwitch,
   MazTabsBar,
   MazTabsContent,
   MazTabsContentItem,
+  MazTextarea,
   MazTransitionExpand,
 }