+++ packages/lib/tests/specs/components/maz-stepper.spec.ts
@@ -1,0 +1,83 @@
+import { mount, VueWrapper } from '@vue/test-utils'
+import MazStepper from '@components/MazStepper.vue'
+import { ComponentPublicInstance } from 'vue'
+
+describe('components/MazStepper.vue', () => {
+  expect(MazStepper).toBeTruthy()
+
+  let wrapper: VueWrapper<ComponentPublicInstance & { [key: string]: any }>
+
+  beforeEach(() => {
+    wrapper = mount(MazStepper, {
+      props: {
+        modelValue: 1,
+        steps: [{ validated: true, disabled: true }],
+      },
+      slots: {
+        'title-1': 'Step 1',
+        'content-1': 'Content 1',
+        'subtitle-2': 'subtitle 2',
+        'content-2': 'Content 2',
+        'title-3': 'Step 3',
+        'content-3': 'Content 3',
+      },
+    })
+  })
+
+  test('Should match with the snapshot', () => {
+    expect(wrapper.html()).toMatchSnapshot()
+  })
+
+  test('Should have the current step according with the model', () => {
+    expect(wrapper.vm.stepCount).toBe(3)
+    expect(wrapper.vm.localModelValue).toBe(1)
+    expect(wrapper.vm.currentStep).toBe(1)
+  })
+
+  test('Should have the current step according with the model', () => {
+    wrapper.vm.selectStep(2)
+    expect(wrapper.vm.localModelValue).toBe(2)
+    expect(wrapper.emitted()['update:model-value']).toStrictEqual([[2]])
+  })
+
+  test('Should returns if title exists', () => {
+    const title1 = wrapper.vm.hasTitleForStep(1)
+    expect(title1).toBeTruthy()
+    const title2 = wrapper.vm.hasTitleForStep(2)
+    expect(title2).toBeFalsy()
+  })
+
+  test('Should returns if subtitle exists', () => {
+    const subtitle1 = wrapper.vm.hasSubtitleForStep(1)
+    expect(subtitle1).toBeFalsy()
+    const subtitle2 = wrapper.vm.hasSubtitleForStep(2)
+    expect(subtitle2).toBeTruthy()
+  })
+
+  test('Should returns if step is validated', () => {
+    const validated1 = wrapper.vm.isStepValidated(1)
+    expect(validated1).toBeTruthy()
+    const validated2 = wrapper.vm.isStepValidated(2)
+    expect(validated2).toBeFalsy()
+  })
+
+  test('Should returns if step is disabled', () => {
+    const disabled1 = wrapper.vm.isStepDisabled(1)
+    expect(disabled1).toBeTruthy()
+    const disabled2 = wrapper.vm.isStepDisabled(2)
+    expect(disabled2).toBeFalsy()
+  })
+
+  test('Should returns if step is the last', () => {
+    const isLast1 = wrapper.vm.isLastStep(1)
+    expect(isLast1).toBeFalsy()
+    const isLast2 = wrapper.vm.isLastStep(3)
+    expect(isLast2).toBeTruthy()
+  })
+
+  test('Should open step if click on it', async () => {
+    const buttons = wrapper.findAll('button')
+    await buttons[1].trigger('click')
+    expect(wrapper.emitted()['update:model-value']).toStrictEqual([[3]])
+  })
+})